requirements.txt
agate==1.9.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
babel==2.16.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
daff==1.3.46
dbt-adapters==1.7.0
dbt-common==1.10.0
dbt-core==1.9.0b2
dbt-extractor==0.5.1
dbt-postgres==1.8.2
dbt-semantic-interfaces==0.7.3
deepdiff==7.0.1
distro==1.9.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib-metadata==6.11.0
isodate==0.6.1
Jinja2==3.1.4
jiter==0.6.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
leather==0.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mashumaro==3.13.1
mdurl==0.1.2
more-itertools==10.5.0
msgpack==1.1.0
narwhals==1.9.2
networkx==3.4.1
numpy==2.1.2
openai==1.51.2
ordered-set==4.1.0
packaging==24.1
pandas==2.2.3
parsedatetime==2.6
pathspec==0.12.1
pillow==10.4.0
protobuf==4.25.5
psycopg2-binary==2.9.9
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowplow-tracker==1.0.3
SQLAlchemy==2.0.35
sqlparse==0.5.1
streamlit==1.39.0
tenacity==9.0.0
text-unidecode==1.3
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
types-requests==2.32.0.20240914
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==5.0.3
zipp==3.20.2
